feat(clerk-js): Redesign Connected Accounts in <UserProfile/>
